pyproject.toml
[project]
name = "template-fastapi"
version = "0.0.1"
description = "A GitHub template repository for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-projects>=1.0.0b12",
    "azure-cosmos>=4.9.0",
    "azure-functions>=1.23.0",
    "azure-identity>=1.23.0",
    "azure-monitor-opentelemetry>=1.6.10",
    "azure-storage-blob>=12.25.1",
    "fastapi-mcp>=0.3.4",
    "fastapi[standard]>=0.115.12",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "msgraph-sdk>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "pydantic-settings>=2.10.1",
    "typer>=0.16.0",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
docs = [
    "mkdocs-material>=9.6.12",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.8.0",
    "jupyterlab>=4.4.2",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "ty>=0.0.1a6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.report]
show_missing = true

[tool.ty]
environment = { python-version = "3.10" }

[tool.ty.rules]
unresolved-attribute = "ignore" # Ignore unresolved attributes in classes
possibly-unbound-attribute = "ignore" # Ignore possibly unbound attributes in classes
unknown-argument = "ignore" # Ignore unknown arguments in function calls
invalid-assignment = "ignore" # Ignore invalid assignments
invalid-argument-type = "ignore" # Ignore invalid argument types
